exercise: worlde clone
Implemented the exercises described in the README. Exercise 3 threw
me for a loop because `guesses` for me was an array of objects (e.g.,
`[{id, value}]`) rather than an array of strings and I wasn't using the
nullish coalescing operator or just the general `?` well. Specifically,
I was mapping using `(char, idx)` and I treated `idx` as the index
of the array of objects rather than the index of the character within
`guesses[idx].value`.